+++ services/mcp/src/hono/tool-executor.ts
@@ -8,7 +8,9 @@ import { toolCallDurationSeconds, toolCallsTotal } from './metrics'
 import type { ToolCatalog } from './tool-catalog'
 
 import type { InstructionsBuilder } from './instructions'
-import type { PreBuiltToolEntry, ResolvedState } from './protocol-types'
+import type { Tool as McpTool } from '@modelcontextprotocol/sdk/types.js'
+
+import type { ResolvedState } from './request-state-resolver'
 
 export class ToolExecutor {
     private readonly catalog: ToolCatalog
@@ -19,7 +21,7 @@ export class ToolExecutor {
         this.instructionsBuilder = instructionsBuilder
     }
 
-    async handleToolsList(state: ResolvedState, props: RequestProperties): Promise<{ tools: PreBuiltToolEntry[] }> {
+    async handleToolsList(state: ResolvedState, props: RequestProperties): Promise<{ tools: McpTool[] }> {
         if (state.useSingleExec) {
             return { tools: [this.instructionsBuilder.buildExecToolEntry(state, props)] }
         }
